+++ OPHD/UI/Core/TextField.cpp
@@ -268,8 +268,8 @@ void TextField::drawCursor()
 			// updateCursor() should be called only on events relating to the cursor so this is temporary.
 			updateCursor();
 			auto& renderer = Utility<Renderer>::get();
-			const auto startPosition = NAS2D::Point{static_cast<float>(mCursorX), mRect.y + FIELD_PADDING};
-			const auto endPosition = NAS2D::Point{static_cast<float>(mCursorX), mRect.y + mRect.height - FIELD_PADDING - 1};
+			const auto startPosition = NAS2D::Point{mCursorX, mRect.y + FIELD_PADDING};
+			const auto endPosition = NAS2D::Point{mCursorX, mRect.y + mRect.height - FIELD_PADDING - 1};
 			renderer.drawLine(startPosition + NAS2D::Vector{1, 1}, endPosition + NAS2D::Vector{1, 1}, NAS2D::Color::Black);
 			renderer.drawLine(startPosition, endPosition, NAS2D::Color::White);
 		}
