+++ OPHD/UI/Core/UIContainer.h
@@ -27,7 +27,7 @@ public:
 	std::vector<Control*> controls() const;
 protected:
 	void visibilityChanged(bool visible) override;
-	void positionChanged(float dX, float dY) override;
+	void positionChanged(int dX, int dY) override;
 
 	virtual void onMouseDown(NAS2D::EventHandler::MouseButton button, int x, int y);
